Separate core and game tests
